client/gen_ai/requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
agno
google-api-python-client
groq
ipywidgets
fastapi
uvicorn
requests
